+++ wai-csrf/test/Main.hs
@@ -6,6 +6,8 @@ import Control.Exception qualified as Ex
 import Control.Monad
 import Control.Monad.IO.Class
 import Data.ByteString qualified as B
+import Data.CaseInsensitive qualified as CI
+import Data.Map.Strict qualified as Map
 import Data.Maybe
 import Data.String
 import Network.HTTP.Types qualified as HT
@@ -19,67 +21,7 @@ main :: IO ()
 main = do
    testToken
    testCookies
-   testReject
    putStrLn "TESTS OK"
-
-testReject :: IO ()
-testReject = do
-   tok1 <- WCC.randomToken
-   let c = WCC.defaultConfig
-       sc1 = WCC.setCookie c tok1
-       app = WCC.middleware c \yt _ respond ->
-         respond $
-            W.responseLBS HT.status200 [] $
-               fromString (show (yt == Just tok1))
-
-   tok2 <- WCC.randomToken
-   let sc2 = WCC.setCookie c tok2
-
-   WT.withSession app do
-      -- Make sure the cookie is set
-      WT.assertNoClientCookieExists "t0-a" c.cookieName
-      WT.setClientCookie sc1
-
-      -- Request succeeds because it is GET
-      sres1 <- WT.request WT.defaultRequest
-      WT.assertStatus 200 sres1
-      WT.assertBody "True" sres1
-      WT.assertClientCookieExists "t0-b" c.cookieName
-
-      -- Request fails because it is POST and there is no CSRF header
-      sres2 <- WT.request WT.defaultRequest{W.requestMethod = HT.methodPost}
-      WT.assertStatus 403 sres2
-      WT.assertBody "CSRF" sres2
-
-      -- Request succeeds because it is POST and there is matching CSRF header
-      sres3 <-
-         WT.request
-            WT.defaultRequest
-               { W.requestMethod = HT.methodPost
-               , W.requestHeaders = [("X-CSRF-TOKEN", WC.setCookieValue sc1)]
-               }
-      WT.assertStatus 200 sres3
-      WT.assertBody "True" sres3
-
-      -- Request succeeds because it is GET. The request header doesn't match,
-      -- but it doesn't matter.
-      sres4 <- do
-         WT.request
-            WT.defaultRequest
-               { W.requestHeaders = [("X-CSRF-TOKEN", WC.setCookieValue sc2)]
-               }
-      WT.assertStatus 200 sres4
-      WT.assertBody "True" sres4
-
-      -- Request fails because it is POST, but there is no matching request header
-      sres5 <- do
-         WT.request
-            WT.defaultRequest
-               { W.requestMethod = HT.methodPost
-               , W.requestHeaders = [("X-CSRF-TOKEN", WC.setCookieValue sc2)]
-               }
-      WT.assertStatus 403 sres5
-      WT.assertBody "CSRF" sres5
 
 testToken :: IO ()
 testToken = do
@@ -100,8 +42,10 @@ testCookies :: IO ()
 testCookies = do
    let c = WCC.defaultConfig
 
-   let fapp1 :: (Maybe WCC.Token -> Maybe (Maybe WCC.Token)) -> W.Application
-       fapp1 = \g -> WCC.middleware c (app1 c g)
+   let fapp1
+         :: (Maybe WCC.Token -> Maybe (Maybe WCC.Token))
+         -> W.Application
+       fapp1 = WCC.middleware c . app1 c
 
    -- keeping cookies untouched
    WT.withSession (fapp1 \_ -> Nothing) do
@@ -130,7 +74,8 @@ testCookies = do
       sres1 <- WT.request WT.defaultRequest
       WT.assertBody "Nothing" sres1
       WT.assertClientCookieExists "t2-b" c.cookieName
-      sres2 <- WT.request WT.defaultRequest
+      Just t1 <- Map.lookup c.cookieName <$> WT.getClientCookies
+      sres2 <- WT.request $ addHeader c (WC.setCookieValue t1) WT.defaultRequest
       WT.assertBody (fromString (show (Just tok1))) sres2
       WT.assertClientCookieExists "t2-c" c.cookieName
       WT.getClientCookies
@@ -140,13 +85,17 @@ testCookies = do
       WT.assertNoClientCookieExists "t3-a" c.cookieName
       WT.modifyClientCookies \_ -> ck0
       WT.assertClientCookieExists "t3-b" c.cookieName
-      sres1 <- WT.request WT.defaultRequest
+      Just t1 <- Map.lookup c.cookieName <$> WT.getClientCookies
+      sres1 <- WT.request $ addHeader c (WC.setCookieValue t1) WT.defaultRequest
       WT.assertBody (fromString (show (Just tok1))) sres1
       WT.assertClientCookieExists "t3-c" c.cookieName
       sres2 <- WT.request WT.defaultRequest
       WT.assertBody "Nothing" sres2
       WT.assertClientCookieExists "t3-d" c.cookieName
       WT.assertClientCookieValue "t3-e" c.cookieName ""
+
+addHeader :: WCC.Config -> B.ByteString -> W.Request -> W.Request
+addHeader c t r = r{W.requestHeaders = [(CI.mk c.headerName, t)]}
 
 app1
    :: WCC.Config